backend/requirements.txt
alembic==1.4.3
asyncpg==0.21.0
click==7.1.2
gino==1.1.0b1
gino-starlette==0.1.1
h11==0.11.0
importlib-metadata==1.7.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.8
typing-extensions==3.7.4.3
uvicorn==0.12.2
zipp==3.3.2
zope.interface==5.1.2
